Fixed bug XOR - NN

+++ python/XOR_NN/Module.py
@@ -8,7 +8,7 @@ class Module:
         self.features = features
 
     def forward(self, features) -> np.ndarray:
-        features = np.array(features, dtype=np.float16)
+        features = np.array(features, dtype=np.float64)
 
 class linear(Module):
     def __init__(self, in_features, out_features) -> None:
@@ -16,11 +16,11 @@ class linear(Module):
         self.in_features = in_features
         self.out_features = out_features
 
-        self.weight = np.zeros([out_features, in_features], dtype=np.float16)
-        self.bias = np.zeros([out_features], dtype=np.float16)
+        self.weight = np.zeros([out_features, in_features], dtype=np.float64)
+        self.bias = np.zeros([out_features], dtype=np.float64)
         
-        self.delta_weight = np.zeros([out_features, in_features], dtype=np.float16)
-        self.delta_bias = np.zeros([out_features], dtype=np.float16)
+        self.delta_weight = np.zeros([out_features, in_features], dtype=np.float64)
+        self.delta_bias = np.zeros([out_features], dtype=np.float64)
 
         for i in range(out_features):
             for j in range(in_features):
@@ -29,7 +29,7 @@ class linear(Module):
 
     def forward(self, features) -> np.ndarray:
         super().forward(features)
-        output_pred = np.zeros([self.out_features], dtype=np.float16)
+        output_pred = np.zeros([self.out_features], dtype=np.float64)
         for i in range(self.out_features):
             for j in range(self.in_features):
                 output_pred[i] += features[j] * self.weight[i,j]
